streamlit-email-inbox/requirements.txt
streamlit==1.32.0
supabase==2.14.0
python-dotenv==1.0.0
pandas==2.1.2
html2text==2024.2.25
streamlit-chat==0.1.1
